add file core
