+++ ethexe/consensus/src/validator/mod.rs
@@ -43,33 +43,35 @@
 use crate::{
     BatchCommitmentValidationReply, ConsensusEvent, ConsensusService, SignedProducerBlock,
     SignedValidationRequest,
-    utils::MultisignedBatchCommitment,
     validator::{
-        coordinator::Coordinator, participant::Participant, producer::Producer,
-        submitter::Submitter, subordinate::Subordinate,
+        coordinator::Coordinator,
+        core::{MiddlewareExt, MiddlewareWrapper, ValidatorCore},
+        participant::Participant,
+        producer::Producer,
+        submitter::Submitter,
+        subordinate::Subordinate,
     },
 };
 use anyhow::Result;
-use async_trait::async_trait;
 use derive_more::{Debug, From};
 use ethexe_common::{Address, SimpleBlockData, ecdsa::PublicKey};
 use ethexe_db::Database;
-use ethexe_ethereum::{Ethereum, middleware::Middleware};
+use ethexe_ethereum::Ethereum;
 use ethexe_signer::Signer;
-use futures::{Stream, lock::Mutex, stream::FusedStream};
+use futures::{Stream, stream::FusedStream};
 use gprimitives::H256;
 use initial::Initial;
 use std::{
     collections::VecDeque,
     fmt,
     pin::Pin,
-    sync::Arc,
     task::{Context, Poll},
     time::Duration,
 };
 use submitter::EthereumCommitter;
 
 mod coordinator;
+mod core;
 mod initial;
 mod participant;
 mod producer;
@@ -140,12 +142,10 @@ impl ValidatorService {
                 db: db.clone(),
                 committer: Box::new(EthereumCommitter { router }),
                 middleware: ethereum.middleware().map(|inner| {
-                    Box::new(MiddlewareWrapper {
-                        inner,
-                        db,
-                        cached_election_result: Arc::new(Mutex::new(None)),
-                    }) as Box<dyn MiddlewareExt>
+                    Box::new(MiddlewareWrapper::new(inner, db)) as Box<dyn MiddlewareExt>
                 }),
+                validate_chain_deepness_limit: MAX_CHAIN_DEEPNESS,
+                chain_deepness_threshold: CHAIN_DEEPNESS_THRESHOLD,
             },
             pending_events: VecDeque::new(),
             output: VecDeque::new(),
@@ -440,55 +440,10 @@ impl DefaultProcessing {
 }
 
 #[derive(Debug)]
-struct ValidatorCore {
-    slot_duration: Duration,
-    signatures_threshold: u64,
-    router_address: Address,
-    pub_key: PublicKey,
-
-    #[debug(skip)]
-    signer: Signer,
-    #[debug(skip)]
-    db: Database,
-    #[debug(skip)]
-    committer: Box<dyn BatchCommitter>,
-    #[debug(skip)]
-    middleware: Option<Box<dyn MiddlewareExt>>,
-}
-
-impl Clone for ValidatorCore {
-    fn clone(&self) -> Self {
-        Self {
-            slot_duration: self.slot_duration,
-            signatures_threshold: self.signatures_threshold,
-            router_address: self.router_address,
-            pub_key: self.pub_key,
-            signer: self.signer.clone(),
-            db: self.db.clone(),
-            committer: self.committer.clone_boxed(),
-            middleware: self.middleware.as_ref().map(|x| x.clone_boxed()),
-        }
-    }
-}
-
-#[derive(Debug)]
 struct ValidatorContext {
+    /// Core validator parameters and utilities.
     core: ValidatorCore,
-    // slot_duration: Duration,
-    // signatures_threshold: u64,
-    // router_address: Address,
-    // pub_key: PublicKey,
 
-    // #[debug(skip)]
-    // signer: Signer,
-    // #[debug(skip)]
-    // db: Database,
-    // #[debug(skip)]
-    // committer: Box<dyn BatchCommitter>,
-    // #[debug(skip)]
-    // middleware: Box<dyn MiddlewareExt>,
-    /// Pending events that are saved for later processing.
-    ///
     /// ## Important
     /// New events are pushed-front, in order to process the most recent event first.
     /// So, actually it is a stack.
@@ -508,83 +463,5 @@ impl ValidatorContext {
 
     pub fn pending(&mut self, event: impl Into<PendingEvent>) {
         self.pending_events.push_front(event.into());
-    }
-}
-
-/// Trait for committing batch commitments to the blockchain.
-#[async_trait]
-pub trait BatchCommitter: Send {
-    /// Creates a boxed clone of the committer.
-    fn clone_boxed(&self) -> Box<dyn BatchCommitter>;
-
-    /// Commits a batch of signed commitments to the blockchain.
-    ///
-    /// # Arguments
-    /// * `batch` - The batch of commitments to commit
-    ///
-    /// # Returns
-    /// The hash of the transaction that was sent to the blockchain
-    async fn commit_batch(self: Box<Self>, batch: MultisignedBatchCommitment) -> Result<H256>;
-}
-
-#[derive(Debug)]
-struct ElectionRequest {
-    #[allow(unused)]
-    at_block_hash: H256,
-    #[allow(unused)]
-    at_timestamp: u64,
-    #[allow(unused)]
-    max_validators: u32,
-}
-
-#[async_trait]
-trait MiddlewareExt: Send {
-    /// Creates a boxed clone.
-    fn clone_boxed(&self) -> Box<dyn MiddlewareExt>;
-
-    /// +_+_+
-    #[allow(unused)]
-    async fn make_election_at(self: Box<Self>, request: ElectionRequest) -> Result<Vec<Address>>;
-}
-
-struct MiddlewareWrapper {
-    inner: Middleware,
-    db: Database,
-    #[allow(clippy::type_complexity)]
-    cached_election_result: Arc<Mutex<Option<(ElectionRequest, Vec<Address>)>>>,
-}
-
-#[async_trait]
-impl MiddlewareExt for MiddlewareWrapper {
-    fn clone_boxed(&self) -> Box<dyn MiddlewareExt> {
-        Box::new(Self {
-            inner: self.inner.clone(),
-            db: self.db.clone(),
-            cached_election_result: self.cached_election_result.clone(),
-        })
-    }
-
-    async fn make_election_at(self: Box<Self>, request: ElectionRequest) -> Result<Vec<Address>> {
-        let mut cached = self.cached_election_result.lock().await;
-
-        if let Some((_cached_request, _cached_result)) = &*cached {
-            // TODO: implement this. If cached_request has same at_timestamp and max_validators and
-            // new request at_block_hash is a successor of cached one, then we can reuse cached.
-            Ok(vec![])
-        } else {
-            log::debug!("Making new election request to rpc: {request:?}");
-
-            let result = self
-                .inner
-                .query()
-                .make_election_at(request.at_timestamp, request.max_validators as u128)
-                .await?;
-
-            let result: Vec<Address> = result.into_iter().collect();
-
-            *cached = Some((request, result.clone()));
-
-            Ok(result)
-        }
     }
 }
